CAPSTONE 2/pyproject.toml
[project]
name = "traffic-sign-classification"
version = "0.1.0"
requires-python = "==3.11.*"
dependencies = [
    "tensorflow-cpu==2.15.0",
    "flask",
    "pillow",
    "numpy",
    "pandas",
    "scikit-learn",
]